Remove certificate junk in gateway. Rely on OS for certs and read them into the server
