+++ gateway/environment_variables.example.env
@@ -7,11 +7,6 @@ MAIL_URL_ADDRESS=https://127.0.0.1:3200
 MEDIA_URL_ADDRESS=https://127.0.0.1:3300
 LOGS_URL_ADDRESS=https://127.0.0.1:3400
 
-# certificates
-CERTS_HOSTNAME=awesome.com
+# CERTIFICATES
 CERTS_CRT_FILEPATH=/usr/local/certs/live/cert.pem
 CERTS_KEY_FILEPATH=/usr/local/certs/live/privkey.pem
-CERTS_COUNTRY_CODE=US
-CERTS_STATE=California
-CERTS_CITY=Berkeley
-CERTS_ORGANIZATION=awesome
