+++ gateway/webapi/server/muxrouter/subdomains/subdomains.go
@@ -2,30 +2,9 @@ package subdomains
 
 import (
 	"net/http"
-
-	// "webapi/server/routes/constants"
 )
 
-const XForwardedProto = "X-Forwarded-Proto"
-const https = "https"
-
 type ProxyMux map[string]http.Handler
-
-func (proxyMux ProxyMux) ServeHTTP(w http.ResponseWriter, r *http.Request) {
-	hostname := r.Host
-	subdomain := getSubdomain(hostname)
-	mux := proxyMux[subdomain]
-
-	if mux != nil {
-		// add header here
-		r.Header.Set("X-Forwarded-Proto", "https")
-		mux.ServeHTTP(w, r)
-		return
-	}
-
-	// reroute to 404
-	http.Error(w, http.StatusText(http.StatusNotFound), http.StatusNotFound)
-}
 
 func getSubdomain(hostname string) string {
 	index, subdomain := getDelimiterIndex(0, hostname, ".")
@@ -49,4 +28,18 @@ func getDelimiterIndex(index int, hostname string, delimiter string) (int, strin
 	}
 
 	return searchIndex, hostname[initialIndex:searchIndex]
+}
+
+func (proxyMux ProxyMux) ServeHTTP(w http.ResponseWriter, r *http.Request) {
+	hostname := r.Host
+	subdomain := getSubdomain(hostname)
+	mux := proxyMux[subdomain]
+
+	if mux != nil {
+		r.Header.Set("X-Forwarded-Proto", "https")
+		mux.ServeHTTP(w, r)
+		return
+	}
+
+	http.Error(w, http.StatusText(http.StatusNotFound), http.StatusNotFound)
 }
